:recycle:[refactor]: api 형식 통일
