+++ src/app/api/upload/portfolio/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { uploadPortfolio } from 'service/uploadRequest';
+import { uploadPortfolio } from 'service/postRequests';
 import { getCookie } from 'utils/cookieUtils';
 
 export async function POST(request: Request) {
